Off page rendering for fitLabel(...) function implementation
